Added generated code: think it won't change for long time. Removed obsolete protocol package. Partial OID and RelativeOid implementation -- YOU'RE WARNED. Should implement all types and add support at compiler for them.

+++ src/org/lastrix/asn1s/schema/TagClass.java
@@ -62,7 +62,7 @@ public enum TagClass {
 	 *
 	 * @throws IllegalArgumentException if code is not 0x00, 0x40 or 0xC0.
 	 */
-	public static TagClass getByCode(final byte code) throws IllegalArgumentException {
+	public static TagClass getByCode(final int code) throws IllegalArgumentException {
 		if (code == 0x00) {
 			return UNIVERSAL;
 		} else if (code == 0x40) {
